+++ .config.sh
@@ -27,7 +27,7 @@ IDM_DEBUG=false
 IMAGE_KEYROCK=letsfiware/fiware-idm:8.4.0
 
 # Docker image for Postfix
-IMAGE_POSTFIX=letsfiware/postfix:0.39.0-next
+IMAGE_POSTFIX=letsfiware/postfix:0.40.0
 
 #
 # Wilma
@@ -363,7 +363,7 @@ NODE_RED_LOGGING_METRICS=
 NODE_RED_LOGGING_AUDIT=
 
 # Docker image for Node-RED
-IMAGE_NODE_RED=letsfiware/node-red:0.39.0-next
+IMAGE_NODE_RED=letsfiware/node-red:0.40.0
 
 #
 # Grafana
@@ -389,7 +389,7 @@ ZEPPELIN=
 ZEPPELIN_DEBUG=
 
 # Docker image for Zeppelin
-IMAGE_ZEPPELIN=letsfiware/zeppelin:0.39.0-next
+IMAGE_ZEPPELIN=letsfiware/zeppelin:0.40.0
 
 #
 # Queryproxy
@@ -401,7 +401,7 @@ QUERYPROXY=
 QUERYPROXY_LOGLEVEL=info
 
 # Docker image for Queryproxy
-IMAGE_QUERYPROXY=letsfiware/queryproxy:0.39.0-next
+IMAGE_QUERYPROXY=letsfiware/queryproxy:0.40.0
 
 #
 # Tokenproxy
@@ -413,7 +413,7 @@ TOKENPROXY_LOGLEVEL=info
 TOKENPROXY_VERBOSE=
 
 # Docker image for Tokenproxy
-IMAGE_TOKENPROXY=letsfiware/tokenproxy:0.39.0-next
+IMAGE_TOKENPROXY=letsfiware/tokenproxy:0.40.0
 
 #
 # Regproxy
@@ -452,7 +452,7 @@ REGPROXY_LOGLEVEL=info
 REGPROXY_VERBOSE=false
 
 # Docker image for Regproxy
-IMAGE_REGPROXY=letsfiware/regproxy:0.39.0-next
+IMAGE_REGPROXY=letsfiware/regproxy:0.40.0
 
 #
 # MongoDB
@@ -488,7 +488,7 @@ IMAGE_POSTGRES=postgres:15
 IMAGE_NGINX=nginx:1.26
 
 # Docker image for pwgen
-IMAGE_PWGEN=letsfiware/pwgen:0.39.0-next
+IMAGE_PWGEN=letsfiware/pwgen:0.40.0
 
 #
 # Firewall (firewalld)